seperated components into jsx files
